requirements.txt
arrow==1.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
circuitbreaker==2.1.3
click==8.0.4
cryptography==44.0.2
idna==3.10
jmespath==0.10.0
oci==2.150.0
oci-cli==3.54.2
prompt-toolkit==3.0.43
pycparser==2.22
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
terminaltables==3.1.10
types-python-dateutil==2.9.0.20241206
urllib3==2.3.0
wcwidth==0.2.13
python-dotenv==1.0.1